Update README.md
I started Machine Learning  2 weeks ago and this is my first ML project. I picked up Red Wine Quality from kaggle.com. The input variable was on based physio chemical test  such as:fixed acidity,citric acid etc, and the output variable is the quality of the wine. The goal was to use machine learning which physio-chemical properties make a wine good. based on 1 to 10, with 10 being the best.